fix breaking change on rusqlite changing RowIndex's signature
RowIndex must have just been an alias over i32, but now it's a trait
implemented on str and usize, so we need to change mentat's internal
type alias for it to a usize.

+++ query-projector/src/lib.rs
@@ -343,7 +343,7 @@ impl QueryResults {
     }
 }
 
-type Index = i32;            // See rusqlite::RowIndex.
+type Index = usize;            // See rusqlite::RowIndex.
 enum TypedIndex {
     Known(Index, ValueTypeTag),
     Unknown(Index, Index),
