+++ query-projector/src/projectors/pull_two_stage.rs
@@ -124,7 +124,7 @@ impl TupleTwoStagePullProjector {
         // There will be at least as many SQL columns as Datalog columns.
         // gte 'cos we might be querying extra columns for ordering.
         // The templates will take care of ignoring columns.
-        assert!(row.column_count() >= self.len as i32);
+        assert!(row.column_count() >= self.len);
         self.templates
             .iter()
             .map(|ti| ti.lookup(&row))
@@ -208,7 +208,7 @@ impl RelTwoStagePullProjector {
         // There will be at least as many SQL columns as Datalog columns.
         // gte 'cos we might be querying extra columns for ordering.
         // The templates will take care of ignoring columns.
-        assert!(row.column_count() >= self.len as i32);
+        assert!(row.column_count() >= self.len);
         let mut count = 0;
         for binding in self.templates
                            .iter()
